# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iRestora POS API",
        "description": "API Documentation untuk Sistem POS F&B iRestora",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "List tables for an outlet",
                "operationId": "ff9d243a2aac0c783b0b20f5bd9d5913",
                "parameters": [
                    {
                        "name": "outlet_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tables"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tables/{tableId}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get table detail (public - for QR code access)",
                "description": "Mengambil detail meja berdasarkan ID atau QR code token. Dipakai Customer App saat scan QR code.",
                "operationId": "0285b49875a8331d235ca235bab0fb05",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "description": "Table UUID atau QR code token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table detail"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan"
                    }
                }
            }
        },
        "/api/tables/{tableId}/close": {
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Tutup meja secara manual (kasir) - meja kembali berstatus available",
                "description": "Hanya bisa dilakukan jika order aktif di meja tsb sudah berstatus paid. Bisa dilakukan kasir manapun yang sedang shift aktif di outlet tersebut (tidak harus kasir yang membuka order). Aksi ini tercatat di audit_logs.",
                "operationId": "cc87338b67843733670b1c3fffa7eb63",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "description": "Catatan opsional, misal 'meja sudah dibersihkan'",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Meja kembali berstatus available"
                    },
                    "404": {
                        "description": "Meja tidak ditemukan"
                    },
                    "422": {
                        "description": "Meja tidak punya order aktif, atau order belum lunas (status bukan paid)"
                    },
                    "409": {
                        "description": "Anda belum membuka shift"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}